+++ R/package-cregg.R
@@ -8,31 +8,40 @@
 #' @param formula A formula specifying a model to be estimated. All variables should be factors.
 #' @param id An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.
 #' @param weights An (optional) RHS formula specifying a variable holding survey weights.
-#' @param estimate A character string specifying an estimate type. Current options are average marginal component effects (or AMCEs, \dQuote{amce}, estimated via \code{\link{amce}}), display frequencies (\dQuote{freq}, estimated via \code{\link{freqs}}), marginal means (or AMMs, \dQuote{mm}, estimated via \code{\link{mm}}), differences in MMs (\dQuote{mm_differences}, via \code{\link{mm_diffs}}), or differences in AMCEs (\dQuote{amce_differences}, via \code{\link{amce_diffs}}). Additional options may be made available in the future.
+#' @param estimate A character string specifying an estimate type. Current options are average marginal component effects (or AMCEs, \dQuote{amce}, estimated via \code{\link{amce}}), display frequencies (\dQuote{frequncies}, estimated via \code{\link{freqs}}), marginal means (or AMMs, \dQuote{mm}, estimated via \code{\link{mm}}), differences in MMs (\dQuote{mm_differences}, via \code{\link{mm_diffs}}), or differences in AMCEs (\dQuote{amce_differences}, via \code{\link{amce_diffs}}). Additional options may be made available in the future. Non-ambiguous abbreviations are allowed.
 #' @param by A formula containing only RHS variables, specifying grouping factors over which to perform estimation.
-#' @param \dots Additional arguments to \code{\link{amce}}, \code{\link{freqs}}, \code{\link{mm}}, or \code{\link{amce_diffs}}.
+#' @param \dots Additional arguments to \code{\link{amce}}, \code{\link{freqs}}, \code{\link{mm}}, \code{\link{mm_diffs}}, or \code{\link{amce_diffs}}.
 #' @author Thomas J. Leeper <thosjleeper@gmail.com>
 #' @details The main function \code{cj} is a convenience function wrapper around the underlying estimation functions that provide for average marginal component effects (AMCEs), by default, via the \code{\link{amce}} function, marginal means (MMs) via the \code{\link{mm}} function, and display frequencies via \code{\link{freqs}} and \code{\link{props}}. Additional estimands may be supported in the future through their own functions and through the \code{cj} interface. Plotting is provided via ggplot2 for both types of estimates.
 #' 
 #' The only additional functionality provided by \code{cj} over the underlying functions is the \code{by} argument, which will perform operations on subsets of \code{data}, returning a single data frame. This can be useful, for example, for evaluating profile spillover effects and subgroup results.
 #' 
+#' Note: cregg currently only works with full factorial conjoint experiments. Designs involving constraints between features may be supported in the future.
+#' 
 #' @examples
 #' \dontrun{
 #' # load data
 #' requireNamespace("ggplot2")
-#' data("hainmueller")
+#' data("immigration")
+#' data("taxes")
 #' 
 #' # calculate MMs
-#' d1 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + 
+#' d1 <- cj(immigration, ChosenImmigrant ~ Gender + Education + 
 #'          LanguageSkills + CountryOfOrigin + Job + JobExperience + 
 #'          JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID,
-#'          estimate = "mm")
-#' 
+#'          estimate = "mm", h0 = 0.5)
 #' # plot MMs
-#' plot(d1)
+#' plot(d1, vline = 0.5)
+#'
+#' # calculate MMs for survey-weighted data
+#' d1 <- cj(taxes, chose_plan ~ taxrate1 + taxrate2 + taxrate3 +
+#'          taxrate4 + taxrate5 + taxrate6 + taxrev, id = ~ ID,
+#'          weights = ~ weight, estimate = "mm", h0 = 0.5)
+#' # plot MMs
+#' plot(d1, vline = 0.5)
 #'
 #' # MMs split by profile number
-#' stacked <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + 
+#' stacked <- cj(immigration, ChosenImmigrant ~ Gender + Education + 
 #'               LanguageSkills + CountryOfOrigin + Job + JobExperience + 
 #'               JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID,
 #'               estimate = "mm", by = ~ contest_no)
@@ -44,7 +53,7 @@
 #' plot(stacked) + ggplot2::facet_wrap(~contest_no, nrow = 1L)
 #' 
 #' # estimate AMCEs
-#' d2 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + 
+#' d2 <- cj(immigration, ChosenImmigrant ~ Gender + Education + 
 #'          LanguageSkills + CountryOfOrigin + Job + JobExperience + 
 #'          JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)
 #' 
@@ -52,70 +61,13 @@
 #' plot(d2)
 #' 
 #' ## subgroup analysis
-#' hainmueller$ethnosplit <- cut(hainmueller$ethnocentrism, 2)
-#' x <- cj(na.omit(hainmueller), ChosenImmigrant ~ Gender + Education + LanguageSkills,
-#'         id = ~ CaseID, estimate = "mm", by = ~ ethnosplit)
+#' immigration$ethnosplit <- cut(immigration$ethnocentrism, 2)
+#' x <- cj(na.omit(immigration), ChosenImmigrant ~ Gender + Education + LanguageSkills,
+#'         id = ~ CaseID, estimate = "mm", h0 = 0.5, by = ~ ethnosplit)
 #' plot(x, group = "ethnosplit", vline = 0.5)
 #' }
-#' @seealso \code{\link{amce}} \code{\link{mm}} \code{\link{freqs}} \code{\link{mm_diffs}} \code{\link{plot.cj_amce}}
+#' @seealso
+#'  Functions: \code{\link{amce}} \code{\link{mm}} \code{\link{freqs}} \code{\link{mm_diffs}} \code{\link{plot.cj_amce}}
+#'  Data: \code{\link{immigration}} \code{\link{taxes}}
 #' @keywords package 
 NULL
-
-#' @rdname cregg
-#' @export
-cj <-
-function(
-  data,
-  formula,
-  id = NULL,
-  weights = NULL,
-  estimate = c("amce", "freqs", "mm", "amce_differences", "mm_differences"),
-  by = NULL,
-  ...
-) {
-    estimate <- match.arg(estimate)
-    
-    if (!is.null(by)) {
-        # get RHS variables, variable labels, and factor levels
-        by_vars <- all.vars(stats::update(by, 0 ~ . ))
-        
-        # amce_diffs handles `by` internally
-        if (estimate == "mm_differences") {
-            return(mm_diffs(data = data, formula = formula, by = by, id = id, weights = weights, ...))
-        } else if (estimate == "amce_differences") {
-            return(amce_diffs(data = data, formula = formula, by = by, id = id, weights = weights, ...))
-        }
-        
-        # split
-        split_df <- split(data, model.frame(by, data = data, na.action = NULL), sep = "***")
-        # get results for subsets
-        BY <- list()
-        for(i in seq_along(split_df)) {
-            BY[[i]] <- switch(estimate, amce = amce, freqs = freqs, mm = mm)(data = split_df[[i]], formula = formula, id = id, weights = weights, ...)
-            BY[[i]][["BY"]] <- i
-        }
-        ## get names of subsets
-        by_vals <- stats::setNames(do.call("rbind.data.frame", strsplit(names(split_df), "***", fixed = TRUE)), by_vars)
-        by_vals[["BY"]] <- factor(seq_len(nrow(by_vals)))
-        # combine back into data frame
-        out <- do.call("rbind", BY)
-        out <- structure(
-          merge(out, by_vals, by = "BY"),
-          class = c(paste0("cj_", estimate), "data.frame"),
-          BY = TRUE
-        )
-        out$BY <- factor(names(split_df)[out$BY])
-    } else {
-        if (estimate %in% c("mm_differences", "amce_differences")) {
-            stop("Argument 'by' is required when estimate %in% c('mm_differences', 'amce_differences')")
-        }
-        by_vars <- NULL
-        out <- switch(estimate,
-                 amce = amce(data = data, formula = formula, id = id, weights = weights, ...),
-                 freqs = freqs(data = data, formula = formula, id = id, weights = weights, ...),
-                 mm = mm(data = data, formula = formula, id = id, weights = weights, ...)
-               )
-    }
-    # return value
-    return(structure(out, by = by_vars))
-}