+++ NEWS.md
@@ -1,3 +1,14 @@
+# cregg 0.1.13
+
+* Added another example datast, `taxes`, from Ballard-Rosa et al. (2016).
+* Renamed `hainmueller` dataset to `immigration`.
+
+# cregg 0.1.12
+
+* Expanded test suite to cover survey-weighted data. Note: `cj_anova()` currently does not work with weighted data due to a bug in `survey::anova.svyglm()`.
+* Cleaned up internal code for consistency.
+* Added 'statistic' column to function outputs.
+
 # cregg 0.1.11
 
 * Variances returend by `amce_diffs()` now respect clustering. (#9)